Fix remove_redundant_results function for var replacement with const
In 7X, redundant Result node in a planned_stmt is being removed
by function remove_redundant_results in orca.c.

When a const expr is created from a var,
consttypmod is set as default value and it cannot replace a var as is.
Fixed the const.consttypmod value before replacing var with const.