Remove unused system cols for ao tables in Orca
In commit
https://github.com/greenplum-db/gpdb/commit/fd426814dec4f0f59aafe7daf15ce905ff2dafdf,
system columns `cmin`, `cmax`, `xmin`, `xmax` are removed from attributes of AO tables,
since they are not used in AO tables.

But Orca was still adding them to AO tables in function
`CTranslatorRelcacheToDXL::AddSystemColumns`. Changes are made so that,
for AO tables, these columns are not added to table descriptor and thus
they won't be considered as AO table's columns during orca planning